+++ packages/pkg-core/src/validation/index.ts
@@ -1,10 +1,11 @@
 import { moonQueryProjects } from './moon-query-projects'
 import * as config from './config'
 
-export type {
-  Config,
-  ConfigConfigured,
-} from './config'
+// export type {
+//   Config,
+//   // ConfigConfigured,
+//   ConfigR,
+// } from './config'
 
 export const validators = {
   moon: {