+++ packages/tests/plugin-dev-tests/etc/config/run-workspace-non-runnable.js
@@ -1,0 +1,11 @@
+// @ts-check
+/**
+ * @typedef {import('@business-as-code/cli').ConfigRun} Config
+ * @type {Config}
+ */
+const config = {
+  query: 'projectType=library || projectType=application',
+  command: ''
+}
+
+module.exports.config = config;