+++ packages/plugin-dev-essentials/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @business-as-code/plugin-dev-essentials
 
+## 0.0.0-latest-20231218214927
+
+### Patch Changes
+
+- Making a release
+- Updated dependencies
+  - @business-as-code/address@0.0.0-latest-20231218214927
+  - @business-as-code/core@0.0.0-latest-20231218214927
+  - @business-as-code/error@0.0.0-latest-20231218214927
+  - @business-as-code/fslib@0.0.0-latest-20231218214927
+
 ## 0.0.0-latest-20231217101715
 
 ### Patch Changes