+++ packages/tests/pkg-tests-verdaccio/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @business-as-code/tests-verdaccio
 
+## 0.0.0-latest-20231218214927
+
+### Patch Changes
+
+- Updated dependencies
+  - @business-as-code/core@0.0.0-latest-20231218214927
+  - @business-as-code/fslib@0.0.0-latest-20231218214927
+
 ## 0.0.0-latest-20231217101715
 
 ### Patch Changes